+++ src/lib/util/dependencies.ts
@@ -7,7 +7,7 @@ export interface INpmPackageBase {
   version: string;
 }
 
-interface INpmPackage extends INpmPackageBase {
+export interface INpmPackage extends INpmPackageBase {
   dependencies: {
     [key: string]: string,
   };
@@ -20,7 +20,7 @@ interface INpmPackageMapPromise {
   [pkgPath: string]: Promise<INpmPackage | null>;
 }
 
-interface INpmPackageMap {
+export interface INpmPackageMap {
   [pkgPath: string]: INpmPackage | null;
 }
 
@@ -208,6 +208,23 @@ export const _findPackage = ({
   return { isFlattened: false, pkgPath: null, pkgObj: null };
 };
 
+enum IFoundMapEntryState {
+  started,
+  recursing,
+  complete,
+}
+
+interface IFoundMapEntry {
+  state: IFoundMapEntryState;
+  dependencies: {
+    [name: string]: IDependencies | null,
+  };
+}
+
+interface IFoundMap {
+  [filePath: string]: IFoundMapEntry;
+}
+
 // - Populates `pkgMap` with installed `package.json`s
 // - Creates a recursive `IDependencies[]` object that later needs to be
 //   flattened and ranges fixed.
@@ -219,7 +236,7 @@ const _recurseDependencies = ({
   pkgsFilter,
 }: {
   filePath: string,
-  foundMap?: { [filePath: string]: { [name: string]: IDependencies | null } },
+  foundMap?: IFoundMap,
   names: string[],
   pkgMap: INpmPackageMap,
   pkgsFilter?: string[],
@@ -232,7 +249,7 @@ const _recurseDependencies = ({
   return names
     .filter(isIncludedPkg)
     // Inflated current level.
-    .map((name): { pkg: IDependencies, pkgNames: string[] } | null => {
+    .map((name): { pkg: IDependencies, pkgNames: string[], found: IFoundMapEntry } | null => {
       // Find actual location.
       const { isFlattened, pkgPath, pkgObj } = _findPackage({ filePath, name, pkgMap });
 
@@ -242,9 +259,12 @@ const _recurseDependencies = ({
       }
 
       // Build and check cache.
-      const found = _foundMap[pkgPath] = _foundMap[pkgPath] || {};
-      if (found[name]) {
-        return { pkg: found[name] as IDependencies, pkgNames: [] };
+      const found: IFoundMapEntry = _foundMap[pkgPath] = _foundMap[pkgPath] || {
+        dependencies: {},
+        state: IFoundMapEntryState.started,
+      } as IFoundMapEntry;
+      if (found.dependencies[name]) {
+        return { pkg: found.dependencies[name] as IDependencies, pkgNames: [], found };
       }
 
       // Start building object.
@@ -258,28 +278,36 @@ const _recurseDependencies = ({
 
       // Add reference to cache.
       if (!isFlattened) {
-        found[name] = pkg;
+        found.dependencies[name] = pkg;
       }
 
       // Get list of package names to recurse.
       // We **don't** traverse devDeps here because shouldn't have with
       // real, installed packages.
+      //
+      // TODO(129): Traverse optionalDependencies too.
+      // https://github.com/FormidableLabs/inspectpack/issues/129
       const pkgNames = Object.keys(pkgObj.dependencies || {});
-      return { pkg, pkgNames };
+      return { pkg, pkgNames, found };
     })
     // Remove empties
     .filter(Boolean)
     // Lazy recurse after all caches have been filled for current level.
     .map((obj) => {
       // TS: Have to cast because boolean filter isn't inferred correctly.
-      const { pkg, pkgNames } = obj as { pkg: IDependencies, pkgNames: string[] };
+      const { pkg, pkgNames, found } = obj as { pkg: IDependencies, pkgNames: string[], found: IFoundMapEntry };
 
       // Only recurse when have dependencies.
       //
       // **Note**: This also serves as a way for found / cached dependency
       // hits to have this mutation step avoided since we manually return
       // `[]` on a cache hit.
-      if (pkgNames.length) {
+      if (found.state === IFoundMapEntryState.started && pkgNames.length) {
+        // Mark state before recursion so that we do only **one** traversal
+        // per unique file path.
+        // https://github.com/FormidableLabs/inspectpack/issues/128
+        found.state = IFoundMapEntryState.recursing;
+
         pkg.dependencies = _recurseDependencies({
           filePath: pkg.filePath,
           foundMap: _foundMap,
@@ -287,6 +315,8 @@ const _recurseDependencies = ({
           pkgMap,
           pkgsFilter,
         });
+
+        found.state = IFoundMapEntryState.complete;
       }
 
       return pkg;