+++ test/setup.ts
@@ -1,6 +1,4 @@
-import { expect, use } from "chai";
+import { use } from "chai";
 import * as sinonChai from "sinon-chai";
 
 use(sinonChai);
-
-global.expect = expect;